requirements.txt
customtkinter>=5.2.0
yt-dlp>=2024.0.0
selenium>=4.15.0
Pillow>=10.0.0
requests>=2.31.0
webdriver-manager>=4.0.0
